module5-task1: разделение на модули;
